Break GitHub kt (#59)
* Pulled apart the github service and commented out some testing temporarily

* Committing some fixes from last night

* Add stack trae to failure logging

* Added stack trace to logger
